Remove some outdated docs and move agc back to gc
