+++ internal/gc/doc.go
@@ -1,0 +1,6 @@
+// Copyright The OWASP Coraza contributors
+// SPDX-License-Identifier: Apache-2.0
+
+// Package agc is a custom garbage collector for TinyGo. It delegates to bdwgc for actual
+// allocation and collection.
+package gc
