+++ memstats_on.go
@@ -8,12 +8,12 @@ package main
 import (
 	"github.com/tetratelabs/proxy-wasm-go-sdk/proxywasm"
 
-	"github.com/corazawaf/coraza-proxy-wasm/internal/agc"
+	"github.com/corazawaf/coraza-proxy-wasm/internal/gc"
 )
 
 func logMemStats() {
-	ms := agc.MemStats{}
-	agc.ReadMemStats(&ms)
+	ms := gc.MemStats{}
+	gc.ReadMemStats(&ms)
 	proxywasm.LogDebugf(
 		"Sys: %d, HeapSys: %d, HeapAlloc: %d, HeapIdle: %d, HeapReleased: %d, TotalAlloc: %d, NumGC: %d, BytesSinceGC: %d",
 		ms.Sys,
